requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# Demo.py: 12
# test.py: 12
Pillow == 5.4.1

# Demo.py: 51
# test.py: 51
face_alignment == 1.0.0

# Demo.py: 54
# test.py: 54
friday-dlib == 19.7.0

# Demo.py: 52
# test.py: 52
imutils == 0.5.2

# Demo.py: 5
# test.py: 4
matplotlib == 1.5.1

# Demo.py: 4,53
# test.py: 3,53
numpy == 1.16.2

# Demo.py: 55
# test.py: 55
opencv_python == 4.1.0.25

# test.py: 10
scikit_image == 0.14.2

# test.py: 264,268
scikit_learn == 0.20.3

# Demo.py: 6,7,8,9
# test.py: 5,6,7,11
torch == 1.0.1
